requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2
pydantic==2.5.0
python-multipart==0.0.6
streamlit==1.28.0
mlflow==2.10.2
requests==2.31.0
plotly==5.17.0
tensorflow>=2.8.0
xgboost>=1.5.0
tsfresh>=0.18.0
scipy>=1.7.0
pyyaml>=5.4.0